Brian

+++ phase4/libuser.c
@@ -26,12 +26,15 @@ int Sleep(int seconds){
     CHECKMODE;
 
     sysArg.number = SYS_SLEEP;
-    sysArg.arg1 = (void *) (intptr_t) seconds;
+    sysArg.arg1 = (void *) (long) seconds;
 
     USLOSS_Syscall(&sysArg);
 
-    return (long) sysArg.arg4;
+    return (int)(long) sysArg.arg4;
 }
+
+
+
 int DiskRead(void *dbuff, int unit, int track, int first, int sectors, int *status){
     return 0;
 }
